ajustado parte de canais, falta ajustar mensagens de retorno

+++ back/src/channel/channel.service.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   ConflictException,
   Injectable,
   UnauthorizedException,
@@ -14,13 +15,15 @@ export class ChannelService {
     private readonly jwtService: JwtService,
   ) {}
 
-  async create(createChatDto: CreateChannelDto, token: any) {
+  async create(createChanneltDto: CreateChannelDto, token: any) {
     const decodeToken = this.jwtService.decode(token);
     const userId = decodeToken.id;
-    const channel = await this.repository.createChannel(createChatDto, token);
-    this.addMember(userId, channel.channel_id, 'Admin', userId);
-    this.addMember(createChatDto.user_id, channel.channel_id, 'Member', token);
-    return 'Canal criado' + userId;
+    const channel = await this.repository.createChannel(
+      createChanneltDto,
+      userId,
+    );
+    this.addMember(userId, channel.channel_id, 'Admin', token);
+    return 'Canal criado';
   }
 
   async addMember(
@@ -29,12 +32,18 @@ export class ChannelService {
     status: string,
     token: any,
   ) {
+    //todo verificar se o usuário existe antes de cadastrar
     const decodeToken = this.jwtService.decode(token);
     const userId = decodeToken.id;
     const isAdmin = await this.repository.checkUser(channel_id, userId);
     const isOwner = await this.repository.checkOwner(channel_id, userId);
     if (isAdmin || isOwner) {
+      const member = await this.repository.checkMember(member_id, channel_id);
+      if (member) {
+        throw new ConflictException('O membro já está no canal');
+      }
       this.repository.addUserToChannel(member_id, channel_id, status);
+      return 'Adicionado com sucesso';
     } else {
       throw new UnauthorizedException('Você não é admin ou owner');
     }
@@ -50,6 +59,10 @@ export class ChannelService {
       member_id,
     );
     if ((isAdmin || isOwner) && memberIsOwner == false) {
+      const member = await this.repository.checkMember(member_id, channel_id);
+      if (!member) {
+        throw new BadRequestException('O membro não está no canal');
+      }
       this.repository.removeMemberChannel(member_id, channel_id);
     } else {
       throw new UnauthorizedException(
@@ -84,8 +97,8 @@ export class ChannelService {
   async leaveChannel(channel_id: number, token: any) {
     const decodeToken = this.jwtService.decode(token);
     const userId = decodeToken.id;
-    const admins = this.repository.checkAdmins(userId, channel_id);
-    if ((await admins).length === 0) {
+    const admins = await this.repository.checkAdmins(userId, channel_id);
+    if (admins.length === 0) {
       throw new ConflictException(
         'Não é possível sair do canal sem outros administradores.',
       );
@@ -105,12 +118,12 @@ export class ChannelService {
     const channel = await this.repository.findChannel(channel_id);
     const decodeToken = this.jwtService.decode(token);
     const userId = decodeToken.id;
-    if (channel.password === password || channel.type === 'Public') {
-      return await this.repository.addUserToChannel(
-        userId,
-        channel_id,
-        'Member',
-      );
+    if (
+      (channel.password === password || channel.type === 'Public') &&
+      channel.type !== 'Restrict'
+    ) {
+      this.repository.addUserToChannel(userId, channel_id, 'Member');
+      return 'Entrou no canal';
     } else {
       throw new UnauthorizedException('Não é possível entrar no canal.');
     }
@@ -123,7 +136,7 @@ export class ChannelService {
     if (isOwner) {
       this.repository.changePassword(channel_id, password);
     } else {
-      throw new UnauthorizedException('Você não é admin ou owner');
+      throw new UnauthorizedException('Você não é owner');
     }
   }
 }
